Etags for emacs
	modified:   .emacs

Added some notes on using etags. Also added (require 'etags-select) (see below).

	new file:   eetags.sh

Small script for creating a TAGS file for fortran and python code.

	new file:   emacs/etags-select.el

This is a nice utility downloaded from the emacs wiki. If one tries to jump to a
tag, but there are multiple definitions in various files, it will open up a
buffer and allow you to select the appropriate one.